iot_simulator/frontend/requirements.txt
streamlit
google-cloud-bigquery
pandas
plotly
db-dtypes
numpy
google-auth
